requirements.txt
flask
flask-sqlalchemy
python-dotenv
mysqlclient
pandas
kaleido
plotly